fix: slow address query and add endpoint for fetching multiple addresses

+++ src/TeddySwap.Common/Models/Response/LeaderboardResponse.cs
@@ -4,7 +4,7 @@ public class LeaderBoardResponse
 {
     public string TestnetAddress { get; init; } = string.Empty;
     public string MainnetAddress { get; set; } = string.Empty;
-    public int Rank { get; set; }
+    public double Rank { get; set; }
     public decimal BaseReward { get; set; }
     public decimal BaseRewardPercentage { get; set; }
     public decimal BonusReward { get; set; }
